Bug fixed and added exclusion of specific probes
Fixed bug in computing the inverse covariance matrix of the Gaussian case. Also, now it is possible to mask specific probes you want to exclude from the likelihood. This for example allows to build likelihoods for cross-correlations alone (you ask for T and E in fields, but then exclude the TT and EE probes).