Added biased fiducial and partial support for correlated Gaussian
Now it is possible to pass a bias file (analogously to the fiducial and noise files) which biases the fiducial side. This allows to include shifts in the power spectra as cause by foregrounds or systematics. Also, I started to implement the correlated Gaussian case, where different multipoles get correlated.